requirements.txt
pandas
pyarrow
apache-airflow-providers-postgres
python-dotenv
sqlalchemy